# GH-44081: [C++][Parquet] Fix reported metrics in parquet-arrow-reader-writer-benchmark