@mapleFU 2024-09-12
Float16Type is not a physical type, just a logical type. But this LGTM