@pitrou 2024-09-12
> Float16Type is not a physical type, just a logical type.

I know, but this was convenient :-)